Apply patch with --force.
We can't interact with /bin/patch running in subprocess so we let
it do the best and solve the problems later in merge-tool.

When /bin/patch finds already applied patch, it asks:
Reversed (or previously applied) patch detected!  Assume -R? [n]
Apply anyway? [n]

With --force patch tries to apply the (even reversed) patch anyway
and meld then resolves the problem nicely - it can see that
the code is already there and doesn't show any conflict.