Assume test suite is run if there is any command
In %check there might be technically any command
to run the test suite. We should not assume there
is only 'make' or 'test'.

So if there is at least one line with any command
not commented out, we assume the test suite is run.

Also directly return the value, don't use variable.

Signed-off-by: Tomas Hozza <thozza@redhat.com>