archive.py: Removed pseudo archive types for .asc and .sig files
Previously these were needed in order for such Source files to be downloaded from URL. The reason was bad logic in the SpecFile class method responsible for downloading Sources, because it was downloading only files with extensions of known archive types. This is not needed any more as all Sources with URL are downloaded.

Signed-off-by: Tomas Hozza <thozza@redhat.com>