Enhance functionality of SpecFile.set_tag()
When updating value of a tag containing multiple instances of the same
macro (while preserving macros), the algorithm simply uses the value
of the last instance. But the values of the instances can actually vary,
depending on context, resuting in expanded value of a tag being wrong.

Adjust the algorithm so it tries to determine the same value for all
the macro instances.

Add test for such case.

Signed-off-by: Nikola Forró <nforro@redhat.com>