Get potentially missing package signing key
Before beta freeze of a Fedora release it can happen that some packages are
already signed with a GPG key of that release, but the key is not yet installed
as part of fedora-repos package.

Try to download and install such key from upstream repository in base
image to prevent installation failures.

Signed-off-by: Nikola Forró <nforro@redhat.com>