Make SpecFile.set_tag() behave more natural
Splitting tokens on separators like '.' or '-' helps the algorithm for
replacing tag values while preserving macros better determine what parts
of the value should be considered a part of a macro.

For instance replacing '0.7.%{?prever}%{?dist}' with '0.8.b5%{?dist}'
lead to '0.%{?prever}%{?dist}' with %prever set to '8.b5'.
Now it leads to '0.8.%{?prever}%{?dist}' with %prever set to 'b5', which
is obviously more accurate.

Signed-off-by: Nikola Forró <nforro@redhat.com>