Rename sources and patches if necessary
Names of local Sources and Patches may change between 2 version (for
example when %{version} macro is used in their names). This could lead
to a failed rebase, SRPM build will fail.

Try to handle the situation by renaming the patches and sources in
rebased-sources directory, changing their paths in specfile and then
reverting the names back.

Signed-off-by: František Nečas <fifinecas@seznam.cz>