Use C locale for updating %changelog
Having the wrong locale running on your machine shouldn't
result in a changelog with invalid syntax or foreign characters.

Example of a bad date here:

https://src.fedoraproject.org/rpms/oci-kvm-hook/pull-request/2#request_diff

Technically setlocale() is not threadsafe. However it is unclear
whether this is a problem for rebase-helper. If it is a problem,
then we must replace strftime() usage in order to fix this problem.

The test aims to be reproducible, despite every system having all
sorts of different locales installed.