Make copr project creation more robust
There was a regression in copr (https://pagure.io/copr/copr/issue/1922)
which caused the error message received when a repository already exists
to change. Rather than checking the error message and ignoring it, check
in advance if the project exists and only try to create a new one if it
does not.

Signed-off-by: František Nečas <fnecas@redhat.com>