Warn the user about a potential rename
In case of renames of files prepended by some directives such as %doc or
%license, we cannot automatically detect the rename. The directive
handles the installation of the file so removing the file makes rpmbuild
or mock not complain but the file is not there if it was renamed. This
can be common in case of renames such as README -> README.md. Warn the
user that this is happening and that it may be necessary to check the
specfile.

Also refactor the method a bit, split it into smaller methods to
decrease nesting blocks and improve readability.

Signed-off-by: František Nečas <fnecas@redhat.com>